bucketlist index out of range fix

+++ Assets/Scripts/Algorithms/ThetaStar.cs
@@ -44,7 +44,7 @@ namespace Pathfinding.Algorithms
             var startCapacity = nodeCount / 100;
             Stack<Vector3> path = new Stack<Vector3>(startCapacity);
 
-            openNodes = new BucketList<Node>((int)distance / 200, Mathf.Min(settings.Heuristic(start.pos, goal.pos) * settings.greediness, settings.CostIncrease(start.pos, goal.pos) * (1 - settings.greediness)));
+            openNodes = new BucketList<Node>(distance / 200, Mathf.Min(settings.Heuristic(start.pos, goal.pos) * settings.greediness, settings.CostIncrease(start.pos, goal.pos) * (1 - settings.greediness)));
             //openNodes = new MinHeap<Node>(startCapacity * 10);
 
             closedNodes = new HashSet<Node>();
